backend/requirements.txt
# FastAPI 生态
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# 数据库
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# LangChain & LLM
langchain>=0.1.4
langchain-openai>=0.0.5
langchain-community>=0.0.10

# YOLO
ultralytics>=8.1.0

# 数据处理
numpy>=1.26.3
pandas>=2.1.4
pillow>=10.2.0
opencv-python>=4.9.0.80

# WebSocket & 实时通信
websocket-client>=1.7.0

# 配置管理
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# 日志
loguru>=0.7.2

# 工具
tqdm>=4.66.1

# 调试
httpx>=0.26.0

# 测试
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
